+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version XXX:
 
 * Fix assert when basic_stream used as underlying of ssl::stream with zero-length write.
 * Add Sec-* HTTP headers.
+* Fix nullptr implicit cast on `fields::set()`.
 
 --------------------------------------------------------------------------------